# Conflicts:
#	Strava API v3/src/javastrava/api/v3/model/StravaActivity.java

+++ Strava API v3/src/javastrava/api/v3/service/UploadService.java
@@ -9,32 +9,35 @@ import javastrava.api.v3.service.exception.UnauthorizedException;
 
 /**
  * <p>
- * Uploading to Strava is an asynchronous process. A file is uploaded using a multipart/form-data POST request which performs initial checks on the data and
- * enqueues the file for processing. The activity will not appear in other API requests until it has finished processing successfully.
+ * Uploading to Strava is an asynchronous process. A file is uploaded using a multipart/form-data POST request which performs
+ * initial checks on the data and enqueues the file for processing. The activity will not appear in other API requests until it has
+ * finished processing successfully.
  * </p>
  *
  * <p>
- * Processing status may be checked by polling Strava. A one-second or longer polling interval is recommended. The mean processing time is currently around 8
- * seconds. Once processing is complete, Strava will respond to polling requests with the activity's ID.
+ * Processing status may be checked by polling Strava. A one-second or longer polling interval is recommended. The mean processing
+ * time is currently around 8 seconds. Once processing is complete, Strava will respond to polling requests with the activity's ID.
  * </p>
  *
  * <p>
- * Errors can occur during the submission or processing steps and may be due to malformed activity data or duplicate data submission.
+ * Errors can occur during the submission or processing steps and may be due to malformed activity data or duplicate data
+ * submission.
  * </p>
  *
  * <p>
- * Strava supports FIT, TCX and GPX file types as described below. New file types are not on the road map. Developers are encouraged to use one of these types
- * as it will also maximize compatibility with other fitness applications.
+ * Strava supports FIT, TCX and GPX file types as described below. New file types are not on the road map. Developers are encouraged
+ * to use one of these types as it will also maximize compatibility with other fitness applications.
  * </p>
  *
  * <p>
- * All files are required to include a time with each trackpoint or record, as defined by the file format. Information such as lat/lng, elevation, heartrate,
- * etc. is optional. Manual creation of activities without a data file is not currently supported by the API.
+ * All files are required to include a time with each trackpoint or record, as defined by the file format. Information such as
+ * lat/lng, elevation, heartrate, etc. is optional. Manual creation of activities without a data file is not currently supported by
+ * the API.
  * </p>
  *
  * <p>
- * If you feel your file is compatible with the standards but is still not uploading to Strava, please verify that it works with other fitness applications
- * before contacting support.
+ * If you feel your file is compatible with the standards but is still not uploading to Strava, please verify that it works with
+ * other fitness applications before contacting support.
  * </p>
  *
  * @author Dan Shannon
@@ -43,8 +46,8 @@ import javastrava.api.v3.service.exception.UnauthorizedException;
 public interface UploadService extends StravaService {
 	/**
 	 * <p>
-	 * Upon upload, Strava will respond with an upload ID. You may use this ID to poll the status of your upload. Strava recommends polling no more than once a
-	 * second. Polling more frequently is unnecessary. The mean processing time is around 8 seconds.
+	 * Upon upload, Strava will respond with an upload ID. You may use this ID to poll the status of your upload. Strava recommends
+	 * polling no more than once a second. Polling more frequently is unnecessary. The mean processing time is around 8 seconds.
 	 * </p>
 	 *
 	 * <p>
@@ -56,14 +59,15 @@ public interface UploadService extends StravaService {
 	 * @param uploadId
 	 *            Upload id originally returned when the upload was done
 	 * @return Upload response indicating current status of the upload
-	 * @throws UnauthorizedException If the authenticated user does not have write access
+	 * @throws UnauthorizedException
+	 *             If the authenticated user does not have write access
 	 */
-	public StravaUploadResponse checkUploadStatus(final Integer uploadId) throws UnauthorizedException;
+	public StravaUploadResponse checkUploadStatus(final Long uploadId) throws UnauthorizedException;
 
 	/**
 	 * <p>
-	 * Upon upload, Strava will respond with an upload ID. You may use this ID to poll the status of your upload. Strava recommends polling no more than once a
-	 * second. Polling more frequently is unnecessary. The mean processing time is around 8 seconds.
+	 * Upon upload, Strava will respond with an upload ID. You may use this ID to poll the status of your upload. Strava recommends
+	 * polling no more than once a second. Polling more frequently is unnecessary. The mean processing time is around 8 seconds.
 	 * </p>
 	 *
 	 * <p>
@@ -75,9 +79,10 @@ public interface UploadService extends StravaService {
 	 * @param uploadId
 	 *            Upload id originally returned when the upload was done
 	 * @return Upload response indicating current status of the upload
-	 * @throws UnauthorizedException If the authenticated user does not have write access
+	 * @throws UnauthorizedException
+	 *             If the authenticated user does not have write access
 	 */
-	public CompletableFuture<StravaUploadResponse> checkUploadStatusAsync(final Integer uploadId) throws UnauthorizedException;
+	public CompletableFuture<StravaUploadResponse> checkUploadStatusAsync(final Long uploadId) throws UnauthorizedException;
 
 	/**
 	 * <p>
@@ -101,23 +106,26 @@ public interface UploadService extends StravaService {
 	 * @param description
 	 *            (Optional)
 	 * @param _private
-	 *            (Optional) set to 1 to mark the resulting activity as private, 'view_private' permissions will be necessary to view the activity
+	 *            (Optional) set to 1 to mark the resulting activity as private, 'view_private' permissions will be necessary to
+	 *            view the activity
 	 * @param trainer
 	 *            (Optional) activities without lat/lng info in the file are auto marked as stationary, set to 1 to force
 	 * @param commute
-	 * 	          (Optional) set to 1 to mark as commute
+	 *            (Optional) set to 1 to mark as commute
 	 * @param dataType
 	 *            possible values: fit, fit.gz, tcx, tcx.gz, gpx, gpx.gz
 	 * @param externalId
 	 *            (Optional) data filename will be used by default but should be a unique identifier
 	 * @param file
 	 *            the actual activity data, if gzipped the data_type must end with .gz
-	 * @return Returns an Upload Status object. This object will include an English language status. If success, it will indicate the data is still processing.
-	 *         If there was an error, it will describe the error, potentially containing HTML. Upon a successful submission the request will return 201 Created.
-	 *         If there was an error the request will return 400 Bad Request.
+	 * @return Returns an Upload Status object. This object will include an English language status. If success, it will indicate
+	 *         the data is still processing. If there was an error, it will describe the error, potentially containing HTML. Upon a
+	 *         successful submission the request will return 201 Created. If there was an error the request will return 400 Bad
+	 *         Request.
 	 */
-	public StravaUploadResponse upload(final StravaActivityType activityType, final String name, final String description, final Boolean _private, final Boolean trainer, final Boolean commute, final String dataType,
-			final String externalId, final File file);
+	public StravaUploadResponse upload(final StravaActivityType activityType, final String name, final String description,
+			final Boolean _private, final Boolean trainer, final Boolean commute, final String dataType, final String externalId,
+			final File file);
 
 	/**
 	 * <p>
@@ -141,21 +149,24 @@ public interface UploadService extends StravaService {
 	 * @param description
 	 *            (Optional)
 	 * @param _private
-	 *            (Optional) set to 1 to mark the resulting activity as private, 'view_private' permissions will be necessary to view the activity
+	 *            (Optional) set to 1 to mark the resulting activity as private, 'view_private' permissions will be necessary to
+	 *            view the activity
 	 * @param trainer
 	 *            (Optional) activities without lat/lng info in the file are auto marked as stationary, set to 1 to force
 	 * @param commute
-	 * 	          (Optional) set to 1 to mark as commute
+	 *            (Optional) set to 1 to mark as commute
 	 * @param dataType
 	 *            possible values: fit, fit.gz, tcx, tcx.gz, gpx, gpx.gz
 	 * @param externalId
 	 *            (Optional) data filename will be used by default but should be a unique identifier
 	 * @param file
 	 *            the actual activity data, if gzipped the data_type must end with .gz
-	 * @return Returns an Upload Status object. This object will include an English language status. If success, it will indicate the data is still processing.
-	 *         If there was an error, it will describe the error, potentially containing HTML. Upon a successful submission the request will return 201 Created.
-	 *         If there was an error the request will return 400 Bad Request.
+	 * @return Returns an Upload Status object. This object will include an English language status. If success, it will indicate
+	 *         the data is still processing. If there was an error, it will describe the error, potentially containing HTML. Upon a
+	 *         successful submission the request will return 201 Created. If there was an error the request will return 400 Bad
+	 *         Request.
 	 */
-	public CompletableFuture<StravaUploadResponse> uploadAsync(final StravaActivityType activityType, final String name, final String description, final Boolean _private, final Boolean trainer, final Boolean commute, final String dataType,
+	public CompletableFuture<StravaUploadResponse> uploadAsync(final StravaActivityType activityType, final String name,
+			final String description, final Boolean _private, final Boolean trainer, final Boolean commute, final String dataType,
 			final String externalId, final File file);
 }
